pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = ""
authors = [
    {name = "rodrigols89",email = "rodrigols89.inbox@gmail.com"}
]
readme = "README.md"


#############################################################
#                   Dependencies Section                    #
#############################################################

requires-python = ">=3.12,<4.0"

dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]

[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.1"
ruff = "^0.14.11"
pytest = "^9.0.2"


#############################################################
#                       Ruff Section                        #
#############################################################
pre-commit = "^4.5.1"
pytest-cov = "^7.0.0"
pytest-django = "^4.11.1"

[tool.ruff]
line-length = 79
exclude = [
    ".*/migrations/.*",
]

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = "double"


#############################################################
#                      Pytest Section                       #
#############################################################

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]

# Ativar automaticamente o pytest-django
django_find_project = true

# Arquivos/códigos que devem ser ignorados pelo pytest
[tool.coverage.run]
omit = [
    "*/tests/*",
    "*test_*.py",
    "*_test.py",
    "*_tests.py",
    "tests.py",
    "*/migrations/*",
    "*/__init__.py",
]

# Arquivos/códigos que devem ser ignorados no relatório de cobertura
[tool.coverage.report]
omit = [
    "*/tests/*",
    "*test_*.py",
    "*_test.py",
    "*_tests.py",
    "tests.py",
    "*/migrations/*",
    "*/__init__.py",
]


#############################################################
#                     Taskipy Section                       #
#############################################################

[tool.taskipy.tasks]
# ------------------------ ( Linting ) ----------------------
pre_lint = 'ruff check --fix'
lint = 'ruff check'
# ------------------------ ( Testing ) ----------------------
test = "docker compose exec -T web pytest -s -x --cov=. -vv"
post_test = 'docker compose exec -T web coverage html'
# ---------------------- ( Pre-Commit ) ---------------------
precommit = 'pre-commit run --all-files'
# -------------- ( General Docker Management ) --------------
start_compose = 'docker compose up -d'
down_compose = 'docker compose down'
restart_compose = 'docker restart $(docker ps -q)'
build_compose = 'docker compose up --build -d'
clean_compose = """
docker stop $(docker ps -aq) 2>/dev/null || true &&
docker rm $(docker ps -aq) 2>/dev/null || true &&
docker rmi -f $(docker images -aq) 2>/dev/null || true &&
docker volume rm $(docker volume ls -q) 2>/dev/null || true &&
docker system prune -a --volumes -f
"""
# --------------------- ( Docker Exec ) ---------------------
opendb = 'docker exec -it postgresql psql -U raguser -d rag_db'
openweb = 'docker exec -it django bash'
opennginx = 'docker exec -it nginx bash'
openredis = 'docker exec -it redis bash'
# ------------------ ( Project Management ) -----------------
exportdev = "poetry export --without-hashes --with dev --format=requirements.txt --output=requirements-dev.txt"
exportprod = "poetry export --without-hashes --format=requirements.txt --output=requirements.txt"
killserver = 'sudo kill -9 $(sudo lsof -t -i:8000)'  # Kill service/port 8000


#############################################################
#                    Build System Section                   #
#############################################################

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
